reorganisation des dossiers dans le parsing
