+++ srcs/parsing/map_file_content_getter/get_file_content.c
@@ -52,60 +52,7 @@ static char	*read_entire_file(t_game_data *game_data)
 	return (file_content);
 }
 
-size_t	add_line_to_splitted(char *file_content, char **splitted_file_content)
-{
-	static	size_t	lines_count = 0;
-	size_t			line_size;
-
-	line_size = 0;
-	while (file_content[line_size] != '\n' && file_content[line_size] != '\0')
-	{
-		++line_size;
-	}
-	splitted_file_content[lines_count] = ft_strndup(file_content, line_size);
-	if (splitted_file_content[lines_count] == NULL)
-	{
-		//exit
-	}
-	lines_count += 1;
-	return (line_size + 1);
-}
-
-static void	split_file_content(t_game_data *game_data, char *file_content)
-{
-	char	**splitted_file_content;
-
-	splitted_file_content = malloc(sizeof(char *) * (game_data->map_file_lines_number + 1));
-	// if (splitted_file_content == NULL)
-	// {
-	// 	//exit
-	// }
-	splitted_file_content[game_data->map_file_lines_number] = NULL;
-	size_t	lines = 1;
-	while (lines < game_data->map_file_lines_number)
-	{
-		file_content += add_line_to_splitted(file_content, splitted_file_content);
-		++lines;
-	}
-	if (*file_content != '\0')
-	{
-		add_line_to_splitted(file_content, splitted_file_content);
-	}
-	game_data->map_file_data = splitted_file_content;
-}
-
-static void	save_file_content(t_game_data *game_data, char *file_content)
-{
-	split_file_content(game_data, file_content);
-	if (game_data->map_file_data == NULL)
-	{
-		ft_dprintf(STDERR_FILENO, "Error.\n"
-			"Failed to save file content. Exit.\n");
-		exit(FAILURE);
-	}
-}
-
-void	extract_file_content(t_game_data *game_data)
+void	get_file_content(t_game_data *game_data)
 {
 	char	*file_content;
