+++ srcs/parsing/map_file_parser/parse_map_file.c
@@ -14,7 +14,7 @@
 
 t_content_status	parse_map_file(t_game_data *game_data)
 {
-	extract_file_content(game_data);
+	get_file_content(game_data);
 	// if (get_textures_and_colors(game_data) == INVALID_CONTENT
 	// 		|| get_map(game_data) == INVALID_CONTENT)
 	// {
